refactor: Split test-kit-only internals into their own export path
index_internals.ts (./internals) is documented as the surface
apify-mcp-server-internal consumes -- but 13 of the symbols widened
onto it for the tarball-packaging fix (0810e63b) are used exclusively
by this package's own tests/test_kit/**, never by the internal repo:
HELPER_TOOLS, MAX_LIMIT_WITH_INPUT_SCHEMA,
SERVER_MODE_AUTO_DETECTION_ENABLED, SKYFIRE_ENABLED_TOOLS,
RESOURCE_MIME_TYPE, CALL_ACTOR_MCP_MISSING_TOOL_NAME_MSG,
toolCategoriesEnabledByDefault, actorRunOutputSchema, the SERVER_MODE/
TelemetryEnv/ToolEntry types, APIFY_ACTOR_RUN_META_KEY, and
AUTO_INJECTED_TOOLS. Confirmed by grepping apify-mcp-server-internal
for each -- zero real hits (two apparent matches were coincidental
name collisions: internal re-declares its own
SERVER_MODE_AUTO_DETECTION_ENABLED deliberately, and imports a
same-named RESOURCE_MIME_TYPE from @modelcontextprotocol/ext-apps/server).

Moved them to a new src/index_internals_test_kit.ts, exported as
./internals/test-kit (mirrors ./internals' extensionless + .js dual
mapping), with a matching tsconfig paths alias so the no-build
type-check job still resolves it. index_internals.ts still imports
HELPER_TOOLS locally to build the deprecated HelperTools alias
internal actually uses -- just no longer re-exports it under the new
name. Updated all 8 tests/test_kit/** call sites that used any of
these 13 symbols to import from the new path instead.

./internals stays exactly what apify-mcp-server-internal consumes;
nothing it imports moved.
