+++ src/index_internals_test_kit.ts
@@ -1,0 +1,31 @@
+/*
+ Internal exports needed only by this package's own tests/test_kit/**, not by
+ apify-mcp-server-internal. Kept separate from index_internals.ts so that surface
+ stays minimal for its actual consumer (see src/AGENTS.md).
+*/
+
+import { HELPER_TOOLS, MAX_LIMIT_WITH_INPUT_SCHEMA, SERVER_MODE_AUTO_DETECTION_ENABLED } from './const.js';
+import { SKYFIRE_ENABLED_TOOLS } from './payments/const.js';
+import { RESOURCE_MIME_TYPE } from './resources/widgets.js';
+import { CALL_ACTOR_MCP_MISSING_TOOL_NAME_MSG } from './tools/actors/call_actor.js';
+import { toolCategoriesEnabledByDefault } from './tools/index.js';
+import { actorRunOutputSchema } from './tools/structured_output_schemas.js';
+import type { SERVER_MODE, TelemetryEnv, ToolEntry } from './types.js';
+import { APIFY_ACTOR_RUN_META_KEY } from './utils/mcp.js';
+import { AUTO_INJECTED_TOOLS } from './utils/tools_loader.js';
+
+export {
+    HELPER_TOOLS,
+    MAX_LIMIT_WITH_INPUT_SCHEMA,
+    SERVER_MODE_AUTO_DETECTION_ENABLED,
+    SKYFIRE_ENABLED_TOOLS,
+    RESOURCE_MIME_TYPE,
+    CALL_ACTOR_MCP_MISSING_TOOL_NAME_MSG,
+    toolCategoriesEnabledByDefault,
+    actorRunOutputSchema,
+    type SERVER_MODE,
+    type TelemetryEnv,
+    type ToolEntry,
+    APIFY_ACTOR_RUN_META_KEY,
+    AUTO_INJECTED_TOOLS,
+};
