+++ tests/test_kit/cases/registration.cases.ts
@@ -6,8 +6,8 @@ import {
     defaults,
     getCategoryTools,
     getExpectedToolNamesByCategories,
-    HELPER_TOOLS,
 } from '@apify/actors-mcp-server/internals.js';
+import { HELPER_TOOLS } from '@apify/actors-mcp-server/internals/test-kit.js';
 
 import {
     ACTOR_NORMAL_MODE,
